MEA Standard Design 64 Electrode
MEA_Standard.GDS: CAD file built in KLayout for the design of a standard centrally oriented 59 electrode MEA. Note that one pin is used for ground electrode. This design requires 75 mm / 3 inch wafer or larger.
MEA_4x4_Standard.GDS: CAD file built in KLayout for the design of a 4x4 array of standard centrally oriented 59 electrode MEAs. Note that one pin is used for ground electrode on each MEA. This design is meant for 150 mm / 6 inch wafer or larger.